requirements.txt
attrs==23.2.0
behave==1.2.6
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.0
distlib==0.3.6
filelock==3.12.2
flake8==7.1.0
Flask==3.0.3
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
numpy==1.25.0
outcome==1.3.0.post0
packaging==24.1
parse==1.20.2
parse-type==0.6.2
platformdirs==3.8.0
pluggy==1.5.0
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyMySQL==1.1.0
PySocks==1.7.1
pytest==8.2.2
pytest-cov==5.0.0
python-dotenv==1.0.1
requests==2.32.3
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.17
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.23.1
webdriver-manager==4.0.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
